refactor: separate a token manager to several classes
separate a token manager to several classes

MODULE-1-PART-3-gallery
