+++ token/handler.js
@@ -3,11 +3,11 @@ import { TokenManager } from "./token.manager.js";
 const manager = new TokenManager();
 
 export function checkToken() {
-    manager.reSetTokenExpireTime();
-    manager.checkTokenExists();
+    manager.check.checkExpireTime();
+    manager.check.checkTokenExists();
 }
 
 export function setToken(token) {
-    manager.saveToken(token);
-    manager.setTokenExpireTime();
+    manager.set.saveToken(token);
+    manager.set.setExpireTime();
 }
