+++ token/token.setting.js
@@ -1,0 +1,14 @@
+export class TokenSetting {
+    saveToken(token){
+        localStorage.setItem('token', token);
+    }
+
+    setExpireTime() {
+        const oneMinuteInMs = 60000;
+        const tenMinutes = oneMinuteInMs * 10;
+
+        const tokenExpireTime = Date.now() + tenMinutes;
+        localStorage.setItem('tokenExpireTime', tokenExpireTime);
+    }
+
+}
